FROMGIT: dt-bindings: pwm-backlight: Move brightness-levels to optional.
The patch 'backlight: pwm_bl: compute brightness of LED linearly to
human eye' introduced a default brightness-levels table that is used
when brightness-levels is not available in the dts. So move
brightness-levels and default-brightness-level to be optional.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Acked-by: Daniel Thompson <daniel.thompson@linaro.org>
Signed-off-by: Lee Jones <lee.jones@linaro.org>

BUG=b:80590153
TEST=Build kernel for Cheza

Signed-off-by: Evan Green <evgreen@chromium.org>
(cherry picked from commit 980d7812fdb483ba055c842d566aa08e76bf795a
 https://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git master)
Change-Id: Ia7365168b605c21b662457e994834ba13584eeba
Reviewed-on: https://chromium-review.googlesource.com/1112419
Reviewed-by: Douglas Anderson <dianders@chromium.org>