PM / devfreq: Add support for QCOM devfreq firmware
The firmware present in some QCOM chipsets offloads the steps necessary for
changing the frequency of some devices (Eg: L3). This driver implements the
devfreq interface for this firmware so that various governors could be used
to scale the frequency of these devices.

Each client (say cluster 0 and cluster 1) that wants to vote for a
particular device's frequency (say, L3 frequency) is represented as a
separate voter device (qcom,devfreq-fw-voter) that's a child of the
firmware device (qcom,devfreq-fw).

Change-Id: I9ba17dd859b11e12882b945a922bc6cbc5bf86d3
[sibis: Fix unable to allocate ftbl-base]
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>
Signed-off-by: Sibi Sankar <sibis@codeaurora.org>