FIXUP: FROMLIST: i2c: i2c-qcom-geni: Add bus driver for the Qualcomm GENI I2C controller
---

This is a revert of the previous FROMLIST patch that we landed in our
tree at CL:944056 and a squash of a pick of a newer version.

---

This bus driver supports the GENI based i2c hardware controller in the
Qualcomm SOCs. The Qualcomm Generic Interface (GENI) is a programmable
module supporting a wide range of serial interfaces including I2C. The
driver supports FIFO mode and DMA mode of transfer and switches modes
dynamically depending on the size of the transfer.

Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>
Signed-off-by: Sagar Dharia <sdharia@codeaurora.org>
Signed-off-by: Girish Mahadevan <girishm@codeaurora.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
(am from https://patchwork.kernel.org/patch/10460815/)

BUG=b:79612241
TEST=i2ctransfer 11 r200@0x9

Change-Id: I5d95046bfaf35d06fe2521a91853c0d6103fdb9b
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1067660
Tested-by: Stephen Boyd <swboyd@chromium.org>