drm/msm/dpu: Remove dpu_mdss_isr when dpu_mdss_destroy is called
The MDSS device is created before the MSM driver attempts to bind the
sub components. If any of the components return -EPROBE_DEFER the MDSS
device is destroyed and tried again later.

If this happens the "dpu_mdss_isr" interrupt created from the DPU MDSS
is not freed when the MDSS device is destroyed and has a risk of
triggering later and hitting a fault by accessing a mmio region that
no longer exists. Even if the interrupt isn't triggered by
accident when the device attempts to reprobe it would error out
when it tries to re-register the interrupt so unconditionally removing
it in the destroy is the right move.

Switch the device managed "dpu_mdss_isr" to be unmanaged and add a
free_irq() in the mdss destroy function.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>