arm64: dts: qcom: sc7180: Increase the number of interconnect cells
Increase the number of interconnect-cells as we can now include
the tag information. So now the consumers can specify the path
tag as an additional argument to the endpoints. For example,
attaching a tag = 1 to the path would like like this:

interconnects = <&aggre1_noc MASTER_SDCC_2 1 &mem_noc SLAVE_EBI1 1>;

Signed-off-by: Sibi Sankar <sibis@codeaurora.org>